De-duplicate list lookup code and test
Moved code round a little so settings were in a sensible place
